Introduced StandardEncoding and its subclasses
StandardEncoding and its subclasses are classes that enables to change the text encoder behavior.
I now added three custom text encoding variations.
- EoTEncoding
- SegmentedEotEncoding
- SegmentedEncoding

This also made a change to the pipeline arguments. Now prompt and negative_prompt moved to the new Encoding classes and removed from the pipeline arguments.

The new way of calling pipeline is as follows.

images = pipe(
    pipeline_type=pipe_type,
    text_input=StandardEncoding(
        prompt=prompt,
        negative_prompt=negative_prompt,
    ),
    guidance_scale=guidance_scale,
    num_inference_steps=num_steps,
  )

The commit still contains some debug output because the change is still under development.