Introduce MLL generation
MLL (multi-level layer) is a tentative name of a new pipeline implementation (that I came up with in just a few seconds).
I will rename it later.

It synthesizes layers with residuals rather than with latents.
It allows to mark part of layers as "additional" so that additional layers don't affect each other.
Example: https://twitter.com/tomo161382/status/1604992539731595264

The code is very rough and messy. I must refactor and clean it up later.